fix(heartbeats): cast period param so Create's INSERT type-checks
heartbeat.Store.Create bound $6 to the integer period_seconds column and
reused it in make_interval(secs => $6), where secs wants double precision.
Postgres can't deduce one type for a parameter used as both, so every
Create failed with 42P08 "inconsistent types deduced for parameter $6".
This only surfaces against a real database — the store tests are
DATABASE_URL-gated, so the no-DB local runs and the stale CI snapshot
both missed it. Cast the make_interval use to ::int so the parameter is
unambiguously integer in both positions. Verified green against a live
Postgres 16.

+++ internal/heartbeat/store.go
@@ -95,7 +95,7 @@ func (s *Store) Create(ctx context.Context, h Heartbeat) (Heartbeat, error) {
 		    (account_id, created_by_user_id, name, slug, ping_token,
 		     period_seconds, grace_seconds, status, expected_by)
 		VALUES ($1, $2, $3, $4, $5, $6, $7, 'new',
-		        now() + make_interval(secs => $6))
+		        now() + make_interval(secs => $6::int))
 		RETURNING id, status, expected_by, created_at
 	`, h.AccountID, h.CreatedByUserID, h.Name, h.Slug, h.PingToken,
 		h.PeriodSeconds, h.GraceSeconds).
